Do not use QueryProcessor multiple times
The QueryProcessor object is non-threadsafe and it should not be reused.
This is been enforced with runtime state checks from stable-2.15 onwards.

Rather than using the same object through multiple requests, the
QueryChanges healthcheck now will be instantiated with a _provider_ of
QueryChanges, so that a new QueryProcessor is used each time the
healthcheck is executed.

Change-Id: Ifa4e621eaf87a409c428a618e2a7f2dcd0bb2ca7