AuthHealthCheck: evict account cache by username
Make sure to trigger the entire auth stack by evicting
the in-memory account cache for the healthcheck user.

Checking the auth by relying on the cache would not highlight
potential problems related to new users trying to sign up or
existing users logging it after over one day of inactivity.

Change-Id: Ied4b77aeaaf3f064f22815c3582713b24f36b942