Normalize response payload returned by replicas
Change [1] introduced the ability for slaves to expose a healthcheck
endpoint, however it introduced a bug whereby the response payload was
wrapped with the http status and its value, rather than the value
itself.

The following payload was returned:

```
{
 "status_code": 200,
 "value": {
   "elapsed": 82,
   "auth": {...},
   "querychanges": {...},
   "jgit": {...},
   ...
 }
}
```

rather than:

```
{
   "elapsed": 82,
   "auth": {...},
   "querychanges": {...},
   "jgit": {...},
   ...
}
```

Make replicas return the same payload response as the masters.

Bug: Issue 13550
[1] I6ab4524dcaa96d8571c3fad2eb03624f00902943

Change-Id: Ia741d7fedc23a29299046cb5224d36d0b36c7d70