Honour plugin name in URI on Gerrit replica
When the healthcheck plugin is running on a Gerrit replica,
it needs to manage the REST-API through a filter because directly
of lack of support for API infrastructure.

However, when the plugin name is amended, the URI matched
healthcheck plugin name hardcoded, without taking into consideration
the actual name of the plugin, which could be different if amended
by the Gerrit admin during setup.

Use the injected @PluginName string into the filter so that it can
respect the same URI as it was running on Gerrit primary node.

Bug: Issue 14370
Change-Id: I7a5b13f3e31b4071ba717bb7786b80668bb23ec5