Missing global metric
Global metrics for success/failure of the checks rely on
the API ones exposed by default by Gerrit.

Since replicas don't have a REST API those metrics are
not exposed.

As part if this change the calculation of response code code
for the HealthCheckStatusEndpoint has been moved to the
GlobalHealthCheck.
Apart for fixing the responsability of this method, it also helps
in the global metric exposure.

Bug: Issue 14381
Change-Id: I586576d0b49da284004da7dbc4fa69ea4db469ef