Make GlobalHealthCheck reentrant
The GlobalHealthCheck is called by multiple threads
and should never be stored in the instance variable
of a singleton.

The only exception is for the shallow storage of the
latest result that is invoked by the metrics collector,
which is well known to be subject to erasure due
to concurrency and may not be 100% accurate.

Change-Id: I17b3913274d83cf29a51a521a26b3c50d60b4658