Plugin reports wrong number of failures
The increment of the metric failure should happen
*when* a new failure is happening and NOT when the metric
is read from a callback.

The metrics are now published when a check is triggered.
Prometheus will just scrape them when needed without retriggering any
calculation of the metrics.

The metrics are now injected in the AbstractHealthCheck, which
takes care of updating them whenever a check runs.

Bug: Issue 14385
Change-Id: I5f149f6893289c2874d125a9c04862d53cacd19b