Add blocked threads check
Introduce check that detects the blocked threads (enabled by default).
In case when it reaches above the configured percent of all threads (by
default 50%) check fails and instance is considered unhealthy.

Check can:
* be disabled with 'healthcheck.blockedthreads.enabled = false'
* have threshold reconfigured with 'healthcheck.blocked.threshold=XX'

Note that one can fine tune the check to certain group of threads by
configuring threshold for threads with name starting wiht prefix.
For instance in case when Gerrit instance should be considered
unhealthy when 33% of SSH-Interactive-Worker threads are 'BLOCKED'
one can configure it with

  [ healthcheck "blockedthreads" ]
    threshold = SSH-Interactive-Threads=33

Note that multiple checks for multiple thread group could be
configured (for details with examples see the config.md file).

Bug: Issue 14401
Change-Id: I7d0c4b70decf49465a7318abe4a555cde7318833