Fix BlockedThreadsCheck failing with un-blocked threads
The introduction of BlockedThreadsCheck introduced a P0
regression because of the accumulatio of the counter of
blocked threads caused any Gerrit instance to eventually
fail once the counter of blocked threads reached the 50%
of the total number.

Remove the @Singleton annotation misplaced on the check
and make sure that similar issues aren't passed unnoticed
by using the actual Guice injection for creating the
BlockedThreadsCheck in tests rather than calling the
constructor directly.

Bug: Issue 14736
Change-Id: I430a7a576f8c41594c59d71cced6d43d7acf985a