Introduce healthcheck extension
Allow plugins to provide their own health check that should participate
in the definition of healthiness of the node.

In order to achieve this, the signature of `AbstractHealthCheck` had to
be changed, so that no implementation had to be injected.  It now takes
the `MetricMaker` interface, rather than an instance of
`HealthCheckMetrics.Factory`, which would require an implementation of
MetricMaker which is not already bound.

This refactoring leverages on change I63c9b7f2a1, which allows gerrit
plugins to expose a new `Gerrit-ApiModule` in the manifest enabling
other plugins to register their extensions.

Bug: Issue 40015586
Change-Id: Ibc822ec3981eb12a29439fba76706c387aede786