Introduce healthcheck extension
Allow plugins to provide their own health check that should participate
in the definition of healthiness of the node.

In order to achieve this, the signature of `AbstractHealthCheck` had to
be changed, so that no implementation had to be injected.
It now takes the `MetricMaker` interface, rather than an instance of
`HealthCheckMetrics.Factory`, which would require an implementation of
MetricMaker which is not already bound.

This refactoring allows exposing a new `Gerrit-Api-Module` in the manifest, which
enables plugins to register their extensions.

Note that this change depends on I63c9b7f2a1aee, which at the time of
writing is not yet merged.

Change-Id: Ibc822ec3981eb12a29439fba76706c387aede786