Add the ability to fail healthcheck by creating a fail file
This will enable the Gerrit Admin to cause healthcheck failures by
simply creating a file at a configurable location.
This is useful when a node needs to be taken  out of the pool of
available nodes, for example while it undergoes maintanence and should
not be serving requests.

This approach is more reactive than disabling/enabling the plugin as
Gerrit only unloads/reloads plugins once every minute. The file
approach gives more immediate feedback as the next call after the
file creation will immediately fail.

Change-Id: I999ca049cff9213d3720a982530a5b03f2d02e44