Add IndexWritable health check
Introduce a new health check to verify that Gerrit can perform writes to
the changes index.

This is done by indexing a change specified in configuration. As the
index step is synchronous we can verify its success by checking whether
an exception was thrown.

The check is done at ChangeIndexer level, meaning it is agnostic of the
index implementation.

Note that this change targets master as it's a breaking change as it
requires new configuration, and healthchecks are enabled by default.

Bug: Issue 40015289
Change-Id: I724f88ebd02613e318a2886161d94a0c573ca344