Listen to OnlineUpgradeListener.onSuccess event
Issue:
During the Gerrit upgrade sometimes there is also an index upgrade
therefore Gerrit starts with the old index version set as ready and
issues an on-line reindex. It finally switches to the new version and
marks it as ready (marking the previous one as not-ready).

Solution:
Upon the Gerrit start detect the index version as it was before (IOW
monitor the old version's lock files). But in addition listen to the
`OnlineUpgradeListener.onSuccess` event and in case when `changes`
on-line reindex gets finished and its version changed switch to monitor
the new changes index version lock files.

The following scenarios were tested:
* gerrit 3.8.3 site was initiated and ~200 changes were created
* gerrit 3.9.1 init was performed and healthcheck plugin was dropped to
  plugins
* gerrit 3.9.1 was started and `healthcheck~status` endpoint was called
  a few times while reindex of changes was performed - all `passed`
* switch to newer changes index version locks was confirmed with the
  following log entry:

    QueryChangesHealthCheck : Changes index healthcheck switched from index version 82 to 84

  but also by removing `write.lock` from old version (healtcheck passed)
  and from new version (healtheck failed) as expected
* calling manual reindex on the same version was ignored by healtcheck
  - again as expected

Bug: Issue 40015289
Change-Id: I6c2f8feebadafeb89f4ac8c17bce4b96ab8a5685