Run checks in parallel
The execution and check of the result of the individual
checks were made in a serial stream processing, which means
that the healthcheck endpoint could have actually failed
even if the individual healthchecks would have all completed
in a timely manner.

Introduce the parallel processing of the stream so that
all checks are running in parallel and have more chances
to be completed within the expected timeout from the
load balancer.

Change-Id: Ibbcbbc429bab8c4694f626478a1e7832eac6f64b