Short-circuit on failed healthchecks
Before the full set of checks was always executed
even if it was clear that it would have failed
because of one of the checks failing.

Avoid running the additional checks when one
of the checks is failing for any reason and
therefore running the full set would not be
useful.

Introduce the new status NOT_RUN that clarifies
that the check was not executed because it would
have not change the overall status.

This change allows a Gerrit to avoid running
extra checks when the system is already struggling.
Having too may healthchecks could actually prevent
the system to become healthy again. Minimizing
the execution of checks on a struggling system
is paramount for allowing it to recover more
quickly.

Change-Id: I885b3df06b60e322f12b77674be02536e8131499