Disable git space check by default
Since the healthcheck is used also in the high-availability config
using a shared filesystem, this check should be disabled by default.
System administrators should evaluate whether to enable this check
based on their system architecture and the plugins in use.

To support this behavior, the `HealthCheck` interface has been extended
with an `isEnabledByDefault()` method. The `HealthCheckConfig` class
now consults this method to determine whether a check should be enabled
if not explicitly configured.

Change-Id: I159b4f6ca1b78047ef1669fd1cb14b987a5db3e9