Throw HealthCheckFailedException upon failures or timeouts
Throw a proper exception whenever a health check fails or does not
complete within the expected time limit.

Previously, the check was returning a 500 status code with a JSON
payload, which Gerrit Response API tolerated; however, since the
introduction of Ib1a259f9c05, the response cannot be a 500.

All APIs are expected to throw a proper exception to notify that the
execution has failed. To keep the response JSON payload, the healthcheck
plugin needs to register an exception hook that recognises its
HealthCheckFailedException and performed a proper formatting of the JSON
payload.

Bug: Issue 469458273
Change-Id: Ic6aca61070457b79c52579284a7931bb6d2963b5